edit returned data, add [GET] /admin/all-user, refactor files and use prettier
