+++ server/routes/indexRoute.js
@@ -4,16 +4,13 @@
 const productRouter = require('./productRoute');
 const userRouter = require('./userRoute');
 const mainRouter = require('./mainRoute');
-const adminRouter = require('./adminRoute');
 const categoryRouter = require('./categoryRoute');
 const cartRouter = require('./cartRoute');
 const orderRouter = require('./orderRoute');
 const paymentController = require('./paymentRoute');
 const discountController = require('./discountRoute');
 
-
 function routesInit(app) {
-  app.use('/admin', adminRouter);
   app.use('/', productRouter);
   app.use('/', userRouter);
   app.use('/', mainRouter);
